tensorflow-lite: package headers
The headers are not packages with tensorflow-lite, making it not easy to
compile any applications using them.

This patch copies the headers from the source and output folder to
the includedir.

The patch was derived from the content of a pre-compiled tflite package[1],
and from compilation error messages that were returned by rpicam-apps
recipe (which uses this recipe as an optional dependency)

[1]: https://github.com/prepkg/tensorflow-lite-raspberrypi/releases/tag/2.20.0

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>